Added create_sync_version script
This script is the first one of a shelve I'm working on, that allows you to set an special version in an HDA, meant to be used specially on utility nodes, I have a lot of HDAs with nested hdas, that doesn't have UI or I use their UI and they are just calculating an attribute, or doing a simple utility I need, and if I improve this nodes, or change something in them I want in all the father nodes, I need to manually update each of the father nodes. With this system, you don't need to do that, just use the sync version in your father hda, and if you modify the sync version it will automatically update in all the hdas that contains this one. Keep in mind that this has their problems too, if you change outputs, attribute names or anything on the sync version that will travel and possibly break your HDAs! So keep that in mind!

Some examples where I'm using this are a library of utility hdas I made for myself, like a min max position calculator, that outputs as a detail attribute all min and max positions from the input, so I don't need to setup this everytime, another hda that cleans both attributes and groups not being used (attributes where all values are 0 get removed) and some quick attributes generators like grows or infection ones, that have an already established nomenclature and structure so I know that even if I change their whole network inside, the output will always be on the same format.

Have fun!